Add clerk role and permissions
In addition to the clerk OG role, add a Drupal-wide clerk role with
permissions for the Board, Board Type and Person content types.

For now, give users with the clerk role the permission to:

* Add new content
* Edit their own content
* Delete their own content

Ultimately we want clerk users to be able to edit/delete any
board, term and person in their city, but I'm not yet sure of the
best place to implement that permission requirement.

Addresses #3483