Remove explicit paths to page for adding boards
Add a step that implements "When I go to add a board" instead of
using 'When I go to "node/add/board"'.  This makes the language
of the steps consistent with the changes in
c86ffbd94fa3eacf25b9a5d6e0485a04c8018b45, makes the tests more
readable, and makes them less brittle if we place the canonical
page for adding things somewhere other than ``node/add/*``.

Addresses #3484