Handle autocomplete field in Mink tests
These scenarios were failing because the person and board
reference fields weren't being set when creating board terms:

* City field of Board Term is automatically populated
* A clerk can create a board term for a board in her city

I found that the following steps correctly select a value using
an autocomplete widget:

* Fill in the matching value
* Wait 1 second
* Press the down key
* Press enter

I added a new step with a signature like this:

    And I select "value" from the "field name" autocomplete field

that encapsulates this logic.

I also added a step that allows users to insert a "wait" in the
test.

Addresses #3484