Add tests for administrators creating people, boards and board terms
Currently the scenario
"An admin creates a board term for a city that they're not a member of"
fails because the filtering of the ``my_boards`` view filters the
options for the ``board`` field by OG group membership. Since the
admin user is not a member of any groups, they don't get any items
in this list. See https://pm.theworkdept.com/issues/3484#note-15
for a description of this issue.

Addresses #3484