Update entityreference view displays to add display for admins
Update the ``my_boards`` view to add an Entity Reference display
that shows all boards.  Update the Board field of the Board Term
content type to use the custom selection plugin and use the new
view to provide the options.

Clone the ``my_boards`` view as ``my_members`` and change its
content type to be "Member".  Use its Entity Reference displays
to select the options for the Member field of Board Term.

Addresses #3484