Add custom theme scaffolding
Add a custom theme, Onboard Theme, a sub-theme of
http://drupal.org/projects/bootstrap/.

The Bootstrap theme requires a more recent version of jQuery than
the one shipped with Drupal.  Manage this dependency and specify
which version of jQuery to use, in a feature, Onboard Theme Helpers.
This feature module can also be a place to put theme utility
functions.

Update the install profile and makefile to download the custom
theme's dependencies and select the theme at install-time.

Addresses #3543