Add media query polyfill for IE8.
Use Respond.js (https://github.com/scottjehl/Respond) as a
polyfill for media queries in IE8.

I originally wanted to load the polyfill with Modernizr, but ran
into some issues with Modernizr.load() similar to the ones
described at https://github.com/Modernizr/Modernizr/issues/729.
Since this is the only Polyfill we're using so far, and it
includes a check that short-circuits for browsers that support
media queries, we'll just include this script directly for now.

Note that Respond.js and other media query polyfills don't work
with stylesheets included via @import, so you need to enable
CSS aggregation to make this work.

Addresses #3551