Make the front page city boxes links
Update the ``Front Page Block`` display of the ``Public: cities`` view
so that the ``Taxonomy term: Name`` field doesn't have any wrapper
elements. That causes the top-level element to be the ``<a>`` element.

Then, rewrite the output of the field to create the elements needed to
do the vertical centering.  In some cases, give elements the class
names of the wrapper elements, for consistency.  In other cases,
tweak the class names.

I'd prefer to do this in code at the template level, but it's
annoying to have to lookup the path/path alias for each row, and
doing that would require adding the tid field to the view output.
The approach I took here seems like the cleanest way.

Addresses #3552